Refactors the update_socrata_table task_groups.
